create professors table
* refactor Students to be used for either students or professors
* update backend to filter correctly
* add link to user profile
